+++ pydfs_lineup_optimizer/utils.py
@@ -1,10 +1,14 @@
-from typing import Dict, Tuple, List, Iterable, Any
-from collections import Counter
+from __future__ import division
+from typing import Dict, Tuple, List, Iterable, Set, Any, DefaultDict, Optional, TYPE_CHECKING
+from collections import Counter, defaultdict
 from difflib import SequenceMatcher
 from itertools import combinations, chain, permutations
-from pydfs_lineup_optimizer.player import Player, LineupPlayer
 from pydfs_lineup_optimizer.settings import LineupPosition
 from pydfs_lineup_optimizer.exceptions import LineupOptimizerException
+
+
+if TYPE_CHECKING:
+    from pydfs_lineup_optimizer.player import Player, LineupPlayer
 
 
 def list_intersection(first_list, second_list):
@@ -28,8 +32,8 @@ def ratio(search_string, possible_match):
     return max([SequenceMatcher(None, search_string, part).ratio() for part in parts])
 
 
-def get_positions_for_optimizer(positions_list, has_multi_positional_players=True):
-    # type: (List[LineupPosition], bool) -> Dict[Tuple[str, ...], int]
+def get_positions_for_optimizer(positions_list, multi_positions_combinations=None):
+    # type: (List[LineupPosition], Set[Tuple[str, ...]]) -> Dict[Tuple[str, ...], int]
     """
     Convert positions list into dict for using in optimizer.
     """
@@ -40,14 +44,21 @@ def get_positions_for_optimizer(positions_list, has_multi_positional_players=Tru
             lambda p: len(p.positions) < len(key) and list_intersection(key, p.positions), positions_list
         )))
         positions[key] = min_value
-    if not has_multi_positional_players:
+    if not multi_positions_combinations:
         return positions
+    #  Create list of required combinations for consistency of multi-positions
+    for i in range(2, len(multi_positions_combinations)):
+        total_combinations = len(multi_positions_combinations)
+        for positions_tuple in combinations(multi_positions_combinations, i):
+            flatten_positions = tuple(sorted(set(chain.from_iterable(positions_tuple))))
+            multi_positions_combinations.add(flatten_positions)
+        if total_combinations == len(multi_positions_combinations):
+            break
+    multi_positions_combinations.update(positions.keys())
     for i in range(2, len(positions)):
         for positions_tuple in combinations(positions_counter.keys(), i):
-            flatten_positions = tuple(sorted(chain.from_iterable(positions_tuple)))
-            if len(flatten_positions) != len(set(flatten_positions)):
-                continue
-            if flatten_positions in positions:
+            flatten_positions = tuple(sorted(set(chain.from_iterable(positions_tuple))))
+            if flatten_positions in positions or flatten_positions not in multi_positions_combinations:
                 continue
             min_value = sum(positions[pos] for pos in positions_tuple)
             positions[flatten_positions] = min_value
@@ -55,14 +66,14 @@ def get_positions_for_optimizer(positions_list, has_multi_positional_players=Tru
 
 
 def link_players_with_positions(players, positions):
-    # type: (List[Player], List[LineupPosition]) -> Dict[Player, LineupPosition]
+    # type: (List['Player'], List[LineupPosition]) -> Dict['Player', LineupPosition]
     """
     This method tries to set positions for given players, and raise error if can't.
     """
     positions = positions[:]
-    single_position_players = []  # type: List[Player]
-    multi_positions_players = []  # type: List[Player]
-    players_with_positions = {}  # type: Dict[Player, LineupPosition]
+    single_position_players = []  # type: List['Player']
+    multi_positions_players = []  # type: List['Player']
+    players_with_positions = {}  # type: Dict['Player', LineupPosition]
     for player in players:
         if len(player.positions) == 1:
             single_position_players.append(player)
@@ -96,7 +107,7 @@ def link_players_with_positions(players, positions):
 
 
 def get_remaining_positions(positions, unswappable_players):
-    # type: (List[LineupPosition], List[LineupPlayer]) -> List[LineupPosition]
+    # type: (List[LineupPosition], List['LineupPlayer']) -> List[LineupPosition]
     """
     Remove unswappable players positions from positions list
     """
@@ -107,3 +118,16 @@ def get_remaining_positions(positions, unswappable_players):
                 positions.remove(position)
                 break
     return positions
+
+
+def get_players_grouped_by_teams(players):
+    # type: (Iterable['Player']) -> DefaultDict[str, List['Player']]
+    players_by_teams = defaultdict(list)  # type: DefaultDict[str, List['Player']]
+    for player in players:
+        players_by_teams[player.team].append(player)
+    return players_by_teams
+
+
+def process_percents(percent):
+    # type: (Optional[float]) -> Optional[float]
+    return percent / 100 if percent and percent > 1 else percent